requirements.txt
anthropic
langchain
langchain-anthropic
langgraph
fastapi
uvicorn
python-dotenv
pydantic
streamlit
faiss-cpu
sentence-transformers
pandas